+++ R/R-yyjson.R
@@ -186,18 +186,26 @@ write_flag <- list(
 #' @param num_specials Should jsong strings 'NA'/'Inf'/'NaN' in a numeric context 
 #'        be converted to the \code{'special'} R numeric values
 #'        \code{NA, Inf, NaN}, or left as a \code{'string'}. Default: 'special'
+#' @param promote_num_to_string Should numeric values be promoted to strings 
+#'        when they occur within an array with other string values?  Default: FALSE 
+#'        means to keep numerics as numeric value and promote the \emph{container} to
+#'        be a \code{list} rather than an atomic vector when types are mixed.  If \code{TRUE}
+#'        then array of mixed string/numeric types will be promoted to all 
+#'        string values and returned as an atonic character vector.  Set this to \code{TRUE}
+#'        if you want to emulate the behaviour of \code{jsonlite::fromJSON()}
 #'
 #' @seealso [read_flag()]
 #' @return Named list of options
 #' @export
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 from_opts <- function(
-    int64             = c('string', 'bit64'),
-    missing_list_elem = c('null', 'na'),
-    vectors_to_df     = TRUE,
-    str_specials      = c('string', 'special'),
-    num_specials      = c('special', 'string'),
-    yyjson_read_flag  = 0L
+    int64                 = c('string', 'bit64'),
+    missing_list_elem     = c('null', 'na'),
+    vectors_to_df         = TRUE,
+    str_specials          = c('string', 'special'),
+    num_specials          = c('special', 'string'),
+    promote_num_to_string = FALSE,
+    yyjson_read_flag      = 0L
 ) {
   
   structure(