+++ tests/testthat/test-json_promotion.R
@@ -1,0 +1,21 @@
+
+
+js <- r"(
+[
+  {"a":1.234560},
+  {"a":2},
+  {"a":"apple"}
+])"
+
+
+test_that("numeric promotion to string instead of list works", {
+
+  res <- from_json_str(js)
+  expect_identical(res$a, list(1.234560, 2L, "apple"))
+  
+  res <- from_json_str(js, promote_num_to_string = FALSE)
+  expect_identical(res$a, list(1.234560, 2L, "apple"))
+
+  res <- from_json_str(js, promote_num_to_string = TRUE) 
+  expect_identical(res$a, c("1.234560", "2", "apple"))
+})